requirements.txt
pygame
qrcode
python-dotenv
requests
